Program development: Renaming and editing files.
