+++ RGR_system_software/Discriminant.c
@@ -1,0 +1,7 @@
+#include "Discriminant.h"
+
+// Функція для обчислення дискримінанту
+double discriminant(double a, double b, double c) {
+    return b*b - 4*a*c;
+}
+
