--- RGR_system_software/Discriminant.cpp
@@ -1,7 +1,0 @@
-#include "Discriminant.h"
-
-// Функція для обчислення дискримінанту
-double discriminant(double a, double b, double c) {
-    return b*b - 4*a*c;
-}
-
