+++ RGR_system_software/Discriminant.h
