--- RGR_system_software/RGR_system_software.pro
@@ -1,24 +1,0 @@
-#-------------------------------------------------
-#
-# Project created by QtCreator 2023-05-17T10:23:49
-#
-#-------------------------------------------------
-
-QT       += core
-
-QT       -= gui
-
-TARGET = RGR_system_software
-CONFIG   += console
-CONFIG   -= app_bundle
-
-TEMPLATE = app
-
-
-SOURCES += main.cpp \
-    Roots.cpp \
-    Discriminant.cpp
-
-HEADERS += \
-    Roots.h \
-    Discriminant.h
